V2.6-重构了工程，使用float代替double体系，优化了弧度转化、随机数获取，添加了飞行尾焰消散效果，转化为了eclipse-maven工程，但暂未解决字体问题

+++ dogfightZ/src/dogfight_Z/Ammo/DecoyMaker.java
@@ -9,38 +9,38 @@ import graphic_Z.Interfaces.ThreeDs;
 
 public class DecoyMaker
 {
-	public double roll_angle_aircraft[];
-	public double speed_aircraft;
+	public float roll_angle_aircraft[];
+	public float speed_aircraft;
 	public PriorityQueue<Dynamic> effects;
 	
 	public DecoyMaker
 	(
 		int  camp,
-		double location[], 
-		double rollAngle_aircraft[],
-		double speedAircraft,
-		double velocity,
+		float location[], 
+		float rollAngle_aircraft[],
+		float speedAircraft,
+		float velocity,
 		int lifeTime, 
-		double density, 
-		double resistanceRate, 
+		float density, 
+		float resistanceRate, 
 		LinkedList<ThreeDs>	              add_que,
 		LinkedList<ListIterator<ThreeDs>> del_que,
 		PriorityQueue<Dynamic> effect
 	)
 	{
-		double roll[] = new double[3];
+		float roll[] = new float[3];
 		roll[2] = 0;
 		effects = effect;
-		roll_angle_aircraft = new double[3];
+		roll_angle_aircraft = new float[3];
 		roll_angle_aircraft[0] = rollAngle_aircraft[0];
 		roll_angle_aircraft[1] = rollAngle_aircraft[1];
 		roll_angle_aircraft[2] = rollAngle_aircraft[2];
 		
 		speed_aircraft = speedAircraft;
 		
-		for(double x=105 ; x<=255 ; x+=1/density) 
+		for(float x=105 ; x<=255 ; x+=1/density) 
 		{
-			for(double y=-30 ; y<0 ; y+=1/density)
+			for(float y=-30 ; y<0 ; y+=1/density)
 			{
 			    roll[0] = x;
 	            roll[1] = y;
@@ -59,13 +59,13 @@ public class DecoyMaker
 	/*
 	public DecoyMaker
 	(
-		double location[], 
-		double rollAngle_aircraft[],
-		double speedAircraft,
-		double velocity,
+		float location[], 
+		float rollAngle_aircraft[],
+		float speedAircraft,
+		float velocity,
 		int    lifeTime, 
-		double density, 
-		double resistanceRate, 
+		float density, 
+		float resistanceRate, 
 		LinkedList<ThreeDs>	   add_que,
 		LinkedList<ListIterator<ThreeDs>> del_que,
 		PriorityQueue<Dynamic> effect
