+++ dogfightZ/src/graphic_Z/Interfaces/ThreeDs.java
@@ -5,11 +5,11 @@ import java.util.ListIterator;
 public interface ThreeDs extends Runnable
 {
 	public boolean  constructWithLine();
-	public double[] getLocation();
-	public double[] getRollAngle();
+	public float[] getLocation();
+	public float[] getRollAngle();
 	public boolean  getVisible();
 	public int		getPointsCount();
-	public double[] getPoint(int index);
+	public float[] getPoint(int index);
 	public void 	go();
 	public char		getSpecialDisplayChar();
 	public void     setIterator(ListIterator<ThreeDs> itr);
